Move Search to Header folder
